backend/Edutech-master/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.3
django-environ==0.11.2
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
engagespot==0.0.2
et-xmlfile==1.1.0
idna==3.7
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.3
Pillow==10.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
razorpay==1.4.2
requests==2.32.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.2.1
